Jira KAN-34=>239 Thana/Upazila Get API Design
