+++ Hrm.Application/Features/Thana_Upazila/Requests/Queries/GetThana_UpazilaRequest.cs
@@ -1,0 +1,14 @@
+﻿using Hrm.Application.DTOs.Thana_Upazila;
+using MediatR;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Hrm.Application.Features.Thana_Upazila.Requests.Queries
+{
+    public class GetThana_UpazilaRequest : IRequest<object>
+    {
+    }
+}
