+++ Hrm.Application/Hrm.Application.csproj
@@ -34,8 +34,6 @@
 	</ItemGroup>
 
 	<ItemGroup>
-	  <Folder Include="Features\Thana_Upazila\Handlers\Queries\" />
-	  <Folder Include="Features\Thana_Upazila\Requests\Queries\" />
 	  <Folder Include="Features\TrainingType\Handlers\Queries\" />
 	  <Folder Include="Features\TrainingType\Requests\Queries\" />
 	</ItemGroup>
